Rename tutor to tutorial_util
