+++ modi/__main__.py
@@ -9,7 +9,7 @@ import modi
 
 from modi.util.debugging_util import Debugger
 from modi.util.usage_instructor import UsageInstructor
-from modi.util.tutor import Tutor
+from modi.util.tutorial_util import Tutor
 from modi.util.inspection_util import Inspector
 
 from modi.util.firmware_updater import STM32FirmwareUpdater
